Remove Machine Names Due to Security Team Advisory (#1788)
* Remove Machine Names Due to Security Team Advisory.
In response to a warning from the security team regarding exposed
machine names in our public Jenkinsfile.migraphxintegration,
I have removed these references to better protect our
infrastructure. This CI script has not been in use for some time,
and there is an open ticket to update it in the future.
https://github.com/ROCm/rocMLIR-internal/issues/1730
When it's updated, we will devise a secure method for handling
machine configurations and validate models across various
architectures.

* Deleted the file from the repository.