Add missing LDS barriers to attention (#1853)
Add missing LDS Barrier to attention

There is a race condition that was reproducible on navi and MI100.
If there is more than one iteration, some threads might be loading
from LDS while others are in the next iteration writing to the same
buffer. We add a barrier to prevent that if the number of iterations
is bigger than one.