Fix incorrect fusion check (#1956)
In AffixTuningParameters we will calculate the GemmFeatures based on the ops in the func, at which point we will then write this as an attribute to the func. The problem with this is that MIGraphX, and other users of rocMLIR, could query isModuleFusible at any point, potentially before AffixTuningParameters has run. This is perfectly legal behavior, but there was a bug in fusionUtils that expected the GemmFeatures to have already been written to the func when isModuleFusible was called.

The fix for this that I came up with was to call rock::getFeatures directly on ReduceOps and teach getFeatures how to handle these ops.

I also added in a new CAPI test as we seemingly did not have one testing fusibility of reduce ops.